Add tests for Registry::sync signal
This tests two scenarios:
- Simply setting a Registry from wl_display, then initializing it, so it
  emits sync
- Using ConnectionThread, EventQueue, and all that in a different
  QThread.

The latter test currently fails due to SocketNotifier apparently not
being thread-safe. This also fails in other tests, and it works as soon
as I comment thread->moveToThread(ConnectionThread), it is unrelated.
Note that in libkscreen, I don't have that issue so far, so it also
works in the threaded scenario, in principle.